change seed
When this R package repository was created, I also changed some variable names and moved the calculation of efficacy boundaries for MB to earlier in the code. This effectively changed the seed of the procedure, since it changed the order of random number generation. In this commit I've moved the order back, which does not change the procedure (except for how it affects the seed), but now matches the seed used in our paper.

The variable renames made when this repository was created were
"efficacy_boundaries" -> "all_efficacy_boundaries"
"efficacy_boundary_list" -> "efficacy_boundary_existing"
I also changed the format of the returned efficacy bound list, to include all_efficacy_boundaries. For MB, this is a high dimensional array.